requirements.txt
pyautogui
pytesseract
pillow
numpy
colorama
